+++ resources/assets/js/stories/JobPosterBuilder/Details.stories.tsx
